Wasn't retrieving runs from PRs from forks, and other various fixes
Replaced the Github API call that gets all the check runs for a repo (`GET /repos/linkerd/linkerd2/actions/runs`) with the one that gets all the check runs for a given workflow (`GET /repos/linkerd/linkerd2/actions/workflows/:workflow_name/runs`) because the former wasn't returning all the runs for PRs from forks.

This allowed simplifying lots of things, as long as we clarify which are
the workflows that we want to take into account.

Also fixed the pagination logic, which was missing some data as well.

And fixed the height of the bottom charts that was broken when there was
a long bar label.

I updated the test data under `cmd/testdata/` with the data from the
latest run.